+++ .note.txt
@@ -1,0 +1,1 @@
+1. In `fedimint-ts/mint.ts` >> ```interface ReissueRequest {notes: OOBNotes}``` not string...
