+++ pkg/fedimint/types/modules/mint.go
@@ -1,71 +1,4 @@
-package fedimint
-
-import "encoding/json"
-
-func (mint *MintModule) Reissue(request ReissueRequest, federationId *string) (*ReissueResponse, error) {
-	resp, err := mint.Client.post("/mint/reissue", request)
-	if err != nil {
-		return nil, err
-	}
-	var reissueResp ReissueResponse
-	err = json.Unmarshal(resp, &reissueResp)
-	if err != nil {
-		return nil, err
-	}
-	return &reissueResp, nil
-}
-
-func (mint *MintModule) Spend(request SpendRequest, federationId *string) (*SpendResponse, error) {
-	resp, err := mint.Client.post("/mint/spend", request)
-	if err != nil {
-		return nil, err
-	}
-	var spendResp SpendResponse
-	err = json.Unmarshal(resp, &spendResp)
-	if err != nil {
-		return nil, err
-	}
-	return &spendResp, nil
-}
-
-func (mint *MintModule) Validate(request ValidateRequest, federationId *string) (*ValidateResponse, error) {
-	resp, err := mint.Client.post("/mint/validate", request)
-	if err != nil {
-		return nil, err
-	}
-	var validateResp ValidateResponse
-	err = json.Unmarshal(resp, &validateResp)
-	if err != nil {
-		return nil, err
-	}
-	return &validateResp, nil
-}
-
-func (mint *MintModule) Split(request SplitRequest) (*SplitResponse, error) {
-	resp, err := mint.Client.post("/mint/split", request)
-	if err != nil {
-		return nil, err
-	}
-	var splitResp SplitResponse
-	err = json.Unmarshal(resp, &splitResp)
-	if err != nil {
-		return nil, err
-	}
-	return &splitResp, nil
-}
-
-func (mint *MintModule) Combine(request CombineRequest) (*CombineResponse, error) {
-	resp, err := mint.Client.post("/mint/combine", request)
-	if err != nil {
-		return nil, err
-	}
-	var combineResp CombineResponse
-	err = json.Unmarshal(resp, &combineResp)
-	if err != nil {
-		return nil, err
-	}
-	return &combineResp, nil
-}
+package modules
 
 type FederationIdPrefix struct {
 	Value [4]byte `json:"value"`
